add students page added to create group
