+++ frontend/src/styles/create_group.css
@@ -1,27 +1,80 @@
 .root {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height:100vh;
-    width: 100%;
-    flex-direction: column;
-    padding: 16px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  padding: 80px 16px 16px;
+ 
+}
 
+.inputField {
+  margin-top: 10px !important;
+  width: 250px;
+  font-size: 32px !important;
+  margin: 0 auto;
+}
+
+.button {
+  align-items: center;
+  margin-top: 15px !important;
+  width: 100px;
+  height: 35px;
+  margin: 0 auto;
+}
+
+/* add-students */
+body {
+  margin: 10px;
+}
+
+form {
+  border: 1px solid black;
+  border-radius: 10px;
+  display: flex;
+  flex-wrap: wrap;
+ 
+  padding: 10px;
+
+  margin-top: 80px;
+}
+
+.inputFieldsContainer {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  flex-wrap: wrap;
+  gap: 20px;
+  /* Adjust the gap size as needed */
+}
+
+.inp {
+  margin-bottom: 20px;
+  width: calc(50% - 10px);
+}
+
+.AddBtn {
+  top: 10px;
+  margin-right: 10px;
+}
+
+.SubBtn {
+  position: absolute;
+  bottom: 0;
+  top: 20px;
+  right: 0;
+}
+
+.AddBtn,
+.SubBtn {
+  width: 100%;
+}
+
+@media (max-width: 768px) {
+  form {
+    flex-direction: column;
   }
-  
-  .inputField {
-    margin-top: 10px !important;
-    width: 250px;
-    font-size: 32px !important;
-    margin: 0 auto;
+
+  .inp {
+    width: 100%;
   }
-  
-  .button {
-    margin-top: 15px !important;
-    margin-left: 5px !important;
-    width: 100px;
-    height: 35px;
-    display: block;
-    margin: 0 auto;
-  }
-  
+}
