+++ frontend/src/styles/navbar.css
@@ -1,4 +1,8 @@
 .nav {
     width: 100% !important;
     right : 5px;
+    position: sticky;
+    top: 0;
+    left:0;
+    z-index: 1;
 }
